Workaround for Kokkos issue 872.
Recent checkouts of Kokkos have an issue when building for Cuda where
the -arch argument is supplied in both CXXFLAGS and LDFLAGS.  This
causes an error with nvcc when linking.  I added a workaround that
strips the extra -arch argument in LDFLAGS.

I also fixed the CMake warning about \1 in the string parsing.