Remove direct includes of Kokkos_Core.hpp
I added a new header called Genten_Kokkos.h that is the only place
Kokkos_Core.hpp should be directly included.  Just include that file
whenever you need to include Kokkos.  I also added the include of that
to Genten_Util.h, so rarely will it be needed to include it.  This is to
make sure Kokkos-related defines are consistently used across the code.