Improvements to Cuda version of MTTKRP kernel for permutation approach
I reimplemented the Cuda version of the MTTKRP kernel for the
permutation approach, along the lines of how I redid the CPU/KNL
version.  They are very similar, but different enough that they need
different implementations (the extra loops needed by the CPU version add
enough registers to reduce performance in the GPU version).  This
version seems to be a little faster than the previous one.