Allow out-of-source or inline builds of Kokkos.
This merges in changes from the kokkos_inline_build branch to optionally
allow building of Kokkos inline with Genten.  Generally this is much
easier than building Kokkos separately.  However there are situations
where one would need an out-of-source build of Kokkos (for example,
linking with Trilinos), so both approaches are supported.  For doing an
out-of-source build, specify KOKKOS_PATH in your configure script as
usual.  For doing an inline build, don't set that variable and unpack
kokkos into the toplevel of the genten directory.  You then specify
several Kokkos configuration variables in your genten config script.