Experimental refactoring of TinyVec and mttkrp_perm() schedule.
Cleaned out unused code in Genten_TinyVec.hpp and unified the static and
dynamic specializations.  Also provide Cuda specializations for 1-4
vector entries per thread forcing the entries to be in register space.
New experimental mttkrp_perm() schedule for R that still needs to be
investigated for performance on KNL and P100.