Make new mttkrp_perm() run schedule permanent.
Removed the old run schedule and made the new one permanent.  Testing on
P100 and K80 shows signficant improvement with no loss in performance on
KNL or HSW.  However looking at the KNL assembly code, the kernel does not
vectorize as well when R is not on of the static sizes used in the
schedule, although this doesn't seem to hurt performance.