Experimental use of reduction buffers in mttkrp().
This is an initial attempt at using reduction buffers instead of atomics
in the base mttkrp algorithm (for all execution spaces except Cuda,
which still uses atomics).  This uses the Kokkos::ScatterView<> class
from KokkosContainers to do this.  However I had to make some
modifications to this class, so I am also including a locally modified
copy of it until the changes can be pushed into Kokkos.  This improves
performance on CPU/KNL of the base mttkrp which have slow atomics.