Unify implementations of mttkrp for Sptensor and Sptensor_perm
This unifies the implementation of mttkrp for Sptensor and Sptensor_perm
where they use the same driver code, just different kernel
specializations.  This also changes mttkrp for Sptensor to use the
TinyVec-style implementation, instead of the scratch arrays.  This is
somewhat faster for larger R on the CPU and in some cases substantially
faster on the GPU.