First cut at GCP-Opt
This is a first implementation of GCP using optimization methods
provided by ROL.  ROL is optional, and provided by an external build of
Trilinos.  Everything seems to work, although the GCP method hasn't been
well tested for anything but Gaussian.  There is much more to do:

* Implement scarcity
* Implement the ROL::Vector directly wrapping a Ktensor, instead of
  copying between the two.
* Implement better kernels for the objective function and gradient.
* Understand numerical differences between CPU and GPU.
* Explore various optimization methods.
* Extend CMake infrastructure to get Kokkos and other settings from the
  Trilinos build.

I started with a separate driver executable for GCP, but opted to rename
the CP-ALS driver to Genten, and support both methods with that single driver
(since the vast majority of the code was the same).