Redesign ROL::Vector implementation to not require copies to/from Ktensor
I actually have two separate ROL::Vector implementations that do this in
different ways store the data as a vector and create a Ktensor from
views of this data when needed, and store the data as a Ktensor and
implement vector operations looping over entries in the Ktensor. The
former seems to be generally faster, but I haven't profiled it in depth
yet (it also has a disadvantage for MTTKRP because the factor matrix
entries won't be padded).

Supporting both vector implementations required some refactoring of
RolObjective and RolBoundConstraint.