Work on scatter-view version of MTTKRP.
ScatterView is an alternative to using atomics on low thread-count
architectures (e.g., CPU or KNL but not GPU) where each thread creates a
copy of the resulting factor matrix and then contributions are reduced
across these copies at the end of MTTKRP.  The problem with this
approach is that performance degrades when the factor matrices get too
large or there are too many threads.  To mitigate this I added a tile
size for the number of columns that are processed in each MTTKRP.  Right
now it is unclear to me what this number should be, so I added command
line options to choose it.  In the future I may be able to figure out
what the number should be based on the factor matrix size, the number of
threads, and the architecture.  I also need to do more tests to see if
this version should replace the default MTTKRP version.