Refactor Sptensor variations and associated MTTKRP algorithms.
After experimenting with the ScatterView version of MTTKRP, I decided to
make that the default implementation since it results in no loss of
performance when using the atomic variant.  To make it easy to select
between atomics, duplication, and permuted MTTKRP variations, I unified
the Sptensor and Sptensor_perm classes and refactored the MTTKRP kernels
to take a parameter that determines which variant to call.  This also
adds another variance I call "single", which is appropriate for a serial
execution space (or a single thread).  I also removed the Sptensor_row
class and associated MTTKRP algorithm, since it was never competitive.
This required touching all of the drivers since command line options
have changed, as well as any test that calls MTTKRP.