Add matlab interface for calling cp-als/gcp.
You can now call Genten's cp-als/gcp functions from matlab using
Sptensor/Ktensor for the tensor_toolbox.  This should make the code much
more user-friendly.  Unfortunately this entails a copy of the tensor
between Matlab and Genten, because of Genten's row-wise storage of the
coordinates.  Perhaps we can add a sparse tensor format in
tensor_toolbox with a compatible layout (and also stores coordinates as
integers) to avoid the copy.