Add option for using hash map in gcp_sgd.
By modifying a #define in Genten_GCP_SGD.cpp, you can replace the
sort/binary search approach for searching for tensor zeros with a hash
map using Kokkos::UnorderedMap.  Currently it requires hard-coding the
tensor dimension to use the tensor indices as keys in the hash map (it
may be possible to fix this in the future).  My testing shows this is
not any faster on the CPU, and only marginally faster on the GPU, than
sort/binary search, so it is disabled by default.