Add two more GCP_SGD implementations.
GCP_SGD3 is similar to GCP_SGD2, but fuses the sub-sampling and
MTTKRP kernels into a single kernel and uses atomics only in MTTKRP.
This actually speeds SGD up quite a bit on architectures with fast
atomics because all modes of the MTTKRP are done at once.

GCP_SGD2_HW is a HogWild! version of GCP_SGD2 with thread parallelism
over epoch iterations.  It is quite a bit faster, but is either unstable
or not working correctly on larger problems.