Consolidate gcp_sgd implementations.
Now have just two, the main gcp_sgd and another one gcp_sgd2 that does
bulk sampling.  That one is different enough that the logic gets pretty
messy if you try and combine it.  It is also not likely to be something
we keep.

Also changed some command line options to use - instead of _.