Allow for multiple tensor dimensions in GCP hash map.
The Kokkos::UnorderedMap used for the hash map requires a POD key type.
Since we need tensor subscripts as keys, this required a hack to use a
compile-time sized array as a key to store the subscripts, which meant
you could only use the hash map if your tensor dimension matched what
the code was compiled for.  I fixed this by creating a hash map
container that contains hash maps for tensors of dimension 3, 4, 5, ...
which forwards the needed methods on to the appropriate map at run time
based on the supplied tensor dimension.  This will work fine as long as
the tensor dimension is one of the finite set in the container.