Make all genten driver options available in matlab interface.
This makes all genten command line options available through the matlab
interface in a general way by converting the mxArray* arguments to
strings and passing them to the usual parsing code.  This required
modifying the parsing code to except options with and without the
leading '--' since that is awkward in the matlab interface.