Add threshold for duplicated mttkrp method.
This adds logic to control when the "duplicated" mttkrp method is used
based on the mode length.  If the mode is too long or the number of
processors is too large (which can be controlled by the
--mttkrp-duplicated-threshold option) duplicated is replaced by atomic.
This allows duplicated to be used for the short modes but not for the
long modes where it tends to perform poorly.