Better handling for indefinite Gram matrix.
If you chose to run GenTen CP-ALS with the symmetric Gram option (the
default on CPU), the code would error out if the Gram matrix was
indefinite, forcing you to restart using teh full Gram option (which
calls and indefinite solver).  Now the code will warn you, and switch
itself to a symmetric, indefinite solver.  Unforunately this doesn't
quite work the GPU however, because it appears the triangular solve
portion of the indefinite solver is not fully implemented.  So for the
GPU, I implemented the necessary code, but put in an error message
telling the user to restart with a full Gram.  In the future, hopefully
this will work on the GPU.  Note however, that the default for the GPU
is to use the full Gram matrix anyway, since it is much faster.