POC of distribution of data into medium grain format.
This is not super optimal, I probably don't want to be asking MPI to
tell me where each element lives for example. Secondly I need a more
flexible format for MPI_Put than a hard coded array, but the pieces are
starting to come together to the point where we can reuse some of the
existing code for the actual maths.

The next step is to write a program that converts the tensor COO files
into a binary format that I can teach MPI_IO how to read.