Made a thing to try and search for faster allreduce grids
It's not exhaustive or super smart, it just tries to take processors
from the dimension with the fastest all reduces and give them to the
dimension with the slowest all reduces.  If it is faster then keep it,
if it is slower then use the previous one.  The algorithm stops as soon
as a grid is slower than the previous one since that means that simply
modifying two dimensions doesn't help. It doesn't rule out the
possibility that some other configuration that is not reachable in a
single step is faster.