Made the SGD methods more comparable
This is hopefully the last update for what I think will be an abstract,
what we want to accomplish here is to show that elastic averaging can be
faster than AllReducing per epoch.  I'm having quite a hard time trying
to find anything that is faster time to solution than ADAM though.