Add distributed CP-ALS
This merges in changes from the mpi-sgd-cpals branch to add distributed
CP-ALS by adding support for the needed allReduces() in various mttkrp,
ktensor, and factor matrix operations.  Also introduced a DistKtensor
similar to the DistSpTensor to support this, although next I think I am
going to go the otherway and remove both of these classes in favor of a
DistTensorContext that manipulates distributed sptensors/ktensors as
non-member data.