Remove remnants of old implementation
GradTensor was a remnant from the previous sparse gcp-opt
implementation (which doesn't work) that reused the mttkrp inner
implementation with a wrapper for the GCP gradient tensor.  That was an
interesting idea and we could potentially do something similar for dense
once we have a better mttkrp for dense tensors.