Propagate extra_data pattern to Sptensor
Also rework the pattern to use std::any directly, instead of
std::shared_ptr<std::any>, which is generally unnecessary.  If a shared
pointer is needed, one could just use that as the type for the extra
data.  This made it easier to distinguish extra data for values and
subscripts in Sptensor.  Also verify the properties in the
pyttb->pygenten tests.