Fix assumptions that ttb_indx == std::size_t
For a newer versions of Matlab, the return type of mxGetUint64s() is not
the same as std::size_t, causing compiler error when enabling Matlab
unless GenTen is configured to use a different type for ttb_indx.  But
changing it to something other than std::size_t was causing some compile
problem which this fixes.

I thought about changing the Matlab interface so that ttb_indx does not
have to match mxGetUint64s(), but that would introduce copies of the
tensor in the interface and would cause performance problems in subtle
ways.