Add new ktensor dist method "broadcast"
Similar to all-reduce, but distributes factor matrix rows in a processor
grid layer amongst all processors in that layer (instead of
replicating).  I called this "broadcast" since I thought I was going to
use that form of communication in the import phase, but I actually use
all-gatherv, so it is something of a misnormer (and there is already
another algorithm called all-gather).