Fix several DistKTensorUpdate approaches when ktensors are padded
For algorithms such as GCP-SGD, GCP-OPT, and CP-OPT, the ktensor that is
passed into the DistKTensorUpdate constructor may have a different
padding than the one that is used during the computation, since the
latter arises from the KokkosVector.  This was not properly handled, and
caused asserts to fail in the import/export.  Now we ensure a ktensor
with the same padding is passed into the constructor, and also creates
an overlap ktensor with consistent padding.