Remove @loader_path from CMAKE_INSTALL_RPATH
Seems to fix an error when installing pygenten through pip and using
GNU compilers instead of the native clang toolchain.  Before the install
was complaining with errors similar to:

-- Installing: /var/folders/fy/q4l_9xc54pz3fy853y4jcq5w000v7d/T/tmpxg3a45tz/wheel/platlib/pygenten/_pygenten.so
  error: /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/install_name_tool: for: /var/folders/fy/q4l_9xc54pz3fy853y4jcq5w000v7d/T/tmpxg3a45tz/wheel/platlib/pygenten/_pygenten.so (for architecture arm64) option "-add_rpath @loader_path" would duplicate path, file already has LC_RPATH for: @loader_path

and so the rpath wasn't being set correctly and you would get errors
when importing pygenten that the genten libraries couldn't be found.
Why it worked with the native compilers and not GNU is beyond me though.
Hopefully this doesn't cause problems on other architectures.